requirements.txt
flask 
tensorflow
numpy 
matplotlib
seaborn 
openai-whisper
flask_session
flask_sqlalchemy